fix statement create return problem

+++ app/controllers/statements_controller.rb
@@ -20,8 +20,8 @@ class StatementsController < ApplicationController
 
   # GET /statements/new
   def new
-    @statement = Statement.new
-    @test_id = params[:test_id]
+    @statement = Statement.new()
+    @test = params[:test_id]
   end
 
   # GET /statements/1/edit
@@ -32,32 +32,37 @@ class StatementsController < ApplicationController
   # POST /statements.json
   def create
     @statement = Statement.new(statement_params)
+    @test = @statement.test.id
 
     path_to_file = "#{Rails.root}/app/assets/images/cert#{@statement.id}.pdf"
-    pdf = CombinePDF.load "#{@statement.test.certificate.path}"
 
-    #a = "Привет".encode("UTF-8")
+    pdf = CombinePDF.load "#{@statement.test.certificate.path}" 
 
-# create a textbox and add it to the existing pdf on page 2 #{@statement.fio} #{@statement.position} 
-    pdf.pages[0].textbox @statement.position, height: 60, width: 450, y: 400, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 16
-    pdf.pages[0].textbox @statement.place, height: 60, width: 450, y: 380, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 16
-  #Суспеснкого районе г . Казани ниже
-    pdf.pages[0].textbox @statement.phone, height: 60, width: 450, y: 360, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 16
-    pdf.pages[0].textbox @statement.name, height: 60, width: 450, y: 320, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 35
-    pdf.pages[0].textbox @statement.fio, height: 60, width: 450, y: 280, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 35
-# output the new pdf which now contains your dynamic data
-    pdf.save path_to_file
-    File.open(path_to_file) do |f|
-      @statement.certificate = f
-    end
-    File.delete(path_to_file) if File.exist?(path_to_file)
+    if !pdf.pages.empty?
+
+          #a = "Привет".encode("UTF-8")
+
+      # create a textbox and add it to the existing pdf on page 2 #{@statement.fio} #{@statement.position} 
+          pdf.pages[0].textbox @statement.position, height: 60, width: 450, y: 400, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 16
+          pdf.pages[0].textbox @statement.place, height: 60, width: 450, y: 380, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 16
+        #Суспеснкого районе г . Казани ниже
+          pdf.pages[0].textbox @statement.phone, height: 60, width: 450, y: 360, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 16
+          pdf.pages[0].textbox @statement.name, height: 60, width: 450, y: 320, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 35
+          pdf.pages[0].textbox @statement.fio, height: 60, width: 450, y: 280, x: 82, :font => :time_new_roman, font_color: [0, 0.2, 0.26274509803], max_font_size: 35
+      # output the new pdf which now contains your dynamic data
+          pdf.save path_to_file
+          File.open(path_to_file) do |f|
+            @statement.certificate = f
+          end
+          File.delete(path_to_file) if File.exist?(path_to_file)
+  end
 
     respond_to do |format|
       if @statement.save
         format.html { redirect_to @statement, notice: 'Statement was successfully created.' }
         format.json { render :show, status: :created, location: @statement }
       else
-        format.html { render :new }
+        format.html { render :new}
         format.json { render json: @statement.errors, status: :unprocessable_entity }
       end
     end
@@ -121,7 +126,7 @@ class StatementsController < ApplicationController
 
     # Only allow a list of trusted parameters through.
     def statement_params
-      params.require(:statement).permit(:email, :fio, :phone, :position, :place, :name, :test_id, :notify, :paid, :certificate, {search: [:paid, 
+      params.require(:statement).permit(:email, :name, :soname, :father_name, :position, :place, :test_id, :notified, :paid, :certificate, {search: [:paid, 
                                                                                                      :email, :fio, :ticket_id,{date: [:start, :end]}]})
     end
 end
